Use string literals for regions recently added to AWS SDK
The reference to `Region0` in
the `awscala` package object will cause any usage of AWScala to throw
`java.lang.NoSuchFieldError: CN_NORTHWEST_1` if the AWS SDK
is not sufficiently up to date.

While it is best practice to keep pace with library updates
and usage of an SDK older than the one used to compile AWScala
can never be guaranteed to work,
there are many practical situations where the user may not have
control of the SDK active in their JDK.
One example is running a Spark application on AWS EMR.
EMR provides an Amazon-controlled installation of the AWS SDK
that is added to Spark's extraClassPath and gets loaded in
preference to any SDK bundled in the user's application jar.

By avoiding direct reference to the more recent region enum values,
we decrease the chance of throwing errors in such situations.